==> code/3_CleanComplaints.ipynb <==
# Cleaning the Chicago Police Use-Of-Force Citizen Complaints Data
#### Mihir Bhaskar
#### 11/23/2021

The following file reads in a raw .csv data file on citizen complaints of police use-of-force sourced from: https://data.cpdp.co/data/bVBkzB/ (accessed on 21st November, 2021). 

It performs the following tasks:
1. Cleans, processes, and merges all the subtables in the data (i.e. on complaints, police officers, and witnesses), keeping only complaints from 2015 onwards
2. Creates and outputs a basic dataset at the tract-level, called 'CleanComplaints', that has all tract IDs in Chicago along with the number of complaints from that tract. It does this by first doing a a spatial merge based on lat-long with the CleanACSFile data outputted from 2_CleanACS to get the tract ID for each complaint, and aggregates data up to the tract level. 
3. Creates and outputs a full granular dataset, at the complaint-officer level, called 'CleanComplaints_FindingLevel.csv'

import numpy as np
import pandas as pd
import geopandas as gpd
from shapely import wkt
from pyprojroot import here
import datetime as dt


# Import data (downloaded from website linked above as Excel file)

## To-do: use Python's openpyxl library to filter the datasets before importing -> improve speed

cmp = pd.read_excel(here('./data/raw/uof_complaints_chicago.xlsx'), sheet_name='Allegations')
witness = pd.read_excel(here('./data/raw/uof_complaints_chicago.xlsx'), sheet_name='Complaining Witnesses')
officer = pd.read_excel(here('./data/raw/uof_complaints_chicago.xlsx'), sheet_name='Officer Profile')

## 1. Cleaning all datasets

### Cleaning main complaints file

# Filtering the data to only include complaints from 2015 onwards

## There are 8 rows where incident date (what we want to filter on) is missing
cmp['IncidentDate'].isnull().sum()

## Drop cases where incidentdate is missing - these are only 8 observations, and there is no other good way
## to tell when a complaint occured. The start date only refers to the start of the investigation, and this could be
## very different from the actual timing of the complaint.
cmp = cmp.dropna(subset=['IncidentDate'])

## Converting incident date to a date variable
cmp['IncidentDate']= pd.to_datetime(cmp['IncidentDate'])

## Keeping only complaints >= 2015
cmp = cmp[cmp['IncidentDate'] >= '2015-01-01']

## Convert CRID to numeric for merging with other datasets
cmp['CRID'] = pd.to_numeric(cmp['CRID'])

## Keep only useful variables
cmp = cmp[['CRID', 'OfficerID', 'Allegation', 'Finding', 'Outcome', 'Beat', 'IncidentDate', 'StartDate', 'EndDate', 'InvestigatorRank',
           'Latitude', 'Longitude']]


### Cleaning officer and witness tables

# Converting appointment date to correct date format
officer['clean_appdate'] = ''

officer.loc[officer['ApptDate'].notna(), 'clean_appdate'] = pd.to_datetime(officer['ApptDate'], unit='d', origin='1899-12-30')

# Create an ordinal variable for officer rank (1 = police officer, 2 = sargeant, etc.)
# Based this on Ranks section of this Wiki page: https://en.wikipedia.org/wiki/Chicago_Police_Department#Ranks

# Drop officers without a rank (45 officers dropped - this also takes care of officers with other pieces of missing information)
officer.dropna(subset=['Rank'], inplace=True)

l = ['Sergeant Of Police', 'Sergeant Per Arbitra', 'Sgt', 'Sgt Assgn Sec Spec']
regstr = '|'.join(l)
officer.loc[officer['Rank'].str.contains(regstr), 'rank_order'] = 2

l = ['Lieutenant Of Police', 'Lt']
regstr = '|'.join(l)
officer.loc[officer['Rank'].str.contains(regstr), 'rank_order'] = 3

officer.loc[officer['Rank'] == 'Captain Of Police', 'rank_order'] = 4

l = ['Commander', 'Cmdr']
regstr = '|'.join(l)
officer.loc[officer['Rank'].str.contains(regstr), 'rank_order'] = 5

officer.loc[officer['Rank'] == 'Dep Chief', 'rank_order'] = 6

officer.loc[officer['Rank'] == 'Chief', 'rank_order'] = 7

l = ['Assistant Superintendent', 'Asst Deputy Sup', 'Deputy Supt.', 'First Deputy Supt.', 'Supt Of Police']
regstr = '|'.join(l)
officer.loc[officer['Rank'].str.contains(regstr), 'rank_order'] = 8

# Put the rest of the officers as rank one
officer['rank_order'] = officer['rank_order'].fillna(1)

# Keeping only relevant info from the table
officer = officer[['OfficerID', 'Gender', 'Race', 'clean_appdate', 'rank_order', 'Age']]

# Renaming columns to keep it ready for the merge
officer.rename(columns={'Gender':'officer_gender', 'Race': 'officer_race', 'rank_order': 'officer_rank_order', 'Age': 'officer_age',
                      'clean_appdate': 'officer_appt_date'}, inplace=True)

# Printing to check
officer.head()

## Dropping witness age because it's missing for all
witness.drop(columns=['Age'], inplace=True)

## Renaming columns in witness table before merge

witness.rename(columns={'Gender': 'witness_gender', 'Race': 'witness_race'}, inplace = True)

witness.head()

### Combine all data into one merged dataset

# Merge with the main table
cmp_full = cmp.merge(officer, how='left', on=['OfficerID'])

cmp_full = cmp_full.merge(witness, how='left', on=['CRID'])

# Drop full duplicates (unsure why these are created - maybe because of duplicates in officer ID table?)
cmp_full.drop_duplicates(inplace=True)

cmp_full.drop_duplicates(subset=['CRID', 'OfficerID'], inplace=True)

print(cmp_full.columns)
cmp_full.head()

## 2. Create a basic dataset with just # complaints per tract

# Drop all irrelevant columns
basic_df = cmp_full[['CRID', 'Latitude', 'Longitude']]

# Drop duplicates in CRID - in the imported data, there is a row for every officer linked with the complaint
# For now, we are only interested in the total # of unique complaints
basic_df.drop_duplicates(subset=['CRID'], inplace=True)

# Checking the quality/missingness of lat-long data
print(basic_df.isnull().sum()) 

# There are 22 missing lat-long values. We cannot merge these with the appropriate census tract,
# so dropping these. 
# To-do code: check to see if address information is present for these missing lat-longs, and whether they can be geo-coded
basic_df = basic_df.dropna()

# Convert the dataframe to a GeoDataFrame (required for spatial merge with census tracts polygons)
basic_df = gpd.GeoDataFrame(basic_df, geometry=gpd.points_from_xy(basic_df.Longitude, basic_df.Latitude), crs='epsg:4326')

# Importing the census tracts data and converting it to a GeoDataFrame
acs = pd.read_csv(here('./data/CleanACSFile.csv'))

# Keeping only relevant info from acs file for the spatial merge
acs = acs[['geo_id', 'geometry']]

acs['geometry'] = acs['geometry'].apply(wkt.loads)
acs = gpd.GeoDataFrame(acs, crs='epsg:4326')

# Doing the spatial merge to assign a tract ID to every complaint
basic_df = gpd.sjoin(basic_df, acs[['geo_id', 'geometry']], how='left')

basic_df.head()

# Checking how many observations did not match to a tract ID
print(basic_df.isnull().sum()) # 20 observations did not find a census tract match

basic_df[basic_df['geo_id'].isnull()]

# Note: manuall checking these lat-longs, it appears that these are valid-latlongs, but fall
# outside of the City of Chicago boundaries. E.g., CRID 1074942 was from Oak Lawn, which falls to the west of the boundaries

# Since these are outside of the focus area of the study (City of Chicago), we are dropping these 20 observations
basic_df.dropna(inplace=True)

# Grouping by the census ID to create a count of crime IDs for each tract
counts = basic_df[['CRID', 'geo_id']].groupby('geo_id').count()

# Merging these counts back with the full dataset of census tract IDs
merged = acs.merge(counts, how='left', on=['geo_id'])

# Replacing missing crime ID count values with 0 (i.e. missing means there were 0 complaints found in that tract)
merged['CRID'] = merged['CRID'].fillna(0)

# Renaming column
merged.rename(columns={'CRID':'complaint_count'}, inplace=True)

merged.head()

# Export .CSV file to be used in other scripts
merged[['geo_id', 'complaint_count']].to_csv(here('./data/CleanComplaints.csv'),
                                            encoding='utf-8', index=False)

## 3. Create and explort a more detailed finding-level dataset 

'''
To-dos:
- create officer vintage variable
- Extract incident year 
'''

# Spatially merging tract-id on these complaints so we can do merges with more features later
cmp_full = cmp_full.merge(basic_df, how='left', on=['CRID'])

# Grouping allegation dummies/categories
## Physical violence (by hand)
cmp_full['physical_violence'] = 0
l = ['Choked', 'Punch', 'Kicked', 'Strike', 'Push', 'Stomp', 'Take Down', 'Violence', 'Physical Contact']
regstr = '|'.join(l)
cmp_full.loc[cmp_full['Allegation'].str.contains(regstr), 'physical_violence'] = 1

## Injury caused?
cmp_full['injury'] = 0
cmp_full.loc[cmp_full['Allegation'].str.contains('Injury'), 'injury'] = 1
cmp_full.loc[cmp_full['Allegation'].str.contains('No Injury'), 'injury'] = 0

## Firearm or weapon involved
cmp_full['weapon'] = 0
l = ['Firearm', 'Taser', 'Weapon', 'Shots']
regstr = '|'.join(l)
cmp_full.loc[cmp_full['Allegation'].str.contains(regstr), 'weapon'] = 1

## On or off-duty
cmp_full['on_duty'] = 0
cmp_full.loc[cmp_full['Allegation'].str.contains('On Duty'), 'on_duty'] = 1

# Creating dummy categories for Gender and for Race (of both witness and officer)
cmp_full['officer_white'] = 0
cmp_full.loc[cmp_full['officer_race'] == 'White', 'officer_white'] = 1

cmp_full['witness_white'] = 0
cmp_full.loc[cmp_full['witness_race'] == 'White', 'witness_white'] = 1

cmp_full['officer_male'] = 0
cmp_full.loc[cmp_full['officer_gender'] == 'M', 'officer_male'] = 1

cmp_full['witness_male'] = 0
cmp_full.loc[cmp_full['witness_gender'] == 'M', 'witness_male'] = 1

# Creating count for total number of previous cases against the officer

# Order by date of complaint
cmp_full.sort_values('IncidentDate', inplace=True)

# Creat a cumulative count for the officer (which should then become the count of previous cases against then)
cmp_full['officer_cmp_count'] = cmp_full.groupby('OfficerID').cumcount()

# Extracting the year of the incident
cmp_full['incident_year'] = cmp_full['IncidentDate'].dt.year

# Create variable on how long the officer had been employed for at the time of the incident
#cmp_full['officer_vintage'] = cmp_full['IncidentDate'] - cmp_full['officer_appt_date'] 


# Drop irrelevant variables to prevent merging issues later
cmp_full.drop(columns=['Latitude_x', 'Latitude_y', 'Longitude_x', 'Longitude_y', 'geometry', 'index_right'], inplace=True)

## Exporting this full dataset to be read into Question 2
cmp_full.to_csv(here('./data/CleanComplaints_FindingLevel.csv'), encoding='utf-8', index=False)


# References

Code sourced:
- https://stackoverflow.com/questions/38454403/convert-excel-style-date-with-pandas
- https://www.geeksforgeeks.org/python-convert-excel-serial-date-to-datetime/


# cmp[['CRID', 'Finding']].groupby('Finding').count()


### Archive code that could be used/repurposed later

# Seeing the proportion of complaints that were sustained

#cmp.drop_duplicates(subset=['CRID'], inplace=True)


## Dropping irrelevant columns
#officers.drop(['OfficerFirst', 'OfficerLast'])

## Renaming variables before merging on CRID
#officers.columns = ['CRID', '']

## Dropping irrelevant columns
#complaints.drop(['OfficeFirst', 'OfficerLast', 'AllegationCode', 'RecommendedFinding', 'RecommendedOutcome',
#                'FinalFinding', 'FinalOutcome', ''], axis=1)

## Drop irrelevant columns, merge the info across the three datasets
#print(complaints.info(), '\n', comp_witness.info(), '\n', officers.info())

#cmp.head()

#cmp[['Beat', 'Location', 'City']].head(100)

#cmp['Diff'] = np.where(cmp['RecommendedFinding'] == cmp['FinalFinding'], 1, 0)

#cmp['Diff'].mean()

## The unique ID here is CRID and officerID
## Drop the unnecessary columns, merge the good columns from the other datasets, then describe the missingness and uniqueness 

#cmp.describe()
#cmp.info()

#cmp.nunique(axis=0)

#cmp[cmp['CRID'] == '1090030']

#cmp['uid'] = cmp['CRID'] + (cmp['OfficerID']).astype(str)

#cmp.nunique(axis=0)


#dups = cmp[cmp.duplicated(['CRID'])].sort_values('CRID')

#dups.head(100)
