==> code/7_CombineData.ipynb <==
# Combining all the individually-cleaned datasets
#### Mihir Bhaskar
#### 11/25/2021

This file merges all the tract-level information created by the previous cleaning files into one consolidated dataset, and exports it as a .csv file. Additionally, at each step we check to make sure the merge produces only 798 rows, the number of tracts with data in our dataset.

import pandas as pd
import numpy as np
from pyprojroot import here

# Importing datasets
acs = pd.read_csv(here('./data/CleanACSFile.csv'))
cmp = pd.read_csv(here('./data/CleanComplaints.csv'))
opendata = pd.read_csv(here('./data/CleanOpenData.csv'))
stops = pd.read_csv(here('./data/CleanStopReports.csv'))
geo = pd.read_csv(here('./data/CleanGeoFile.csv'))

# Merging them- each step produces 798 rows. Merges succeed!
final = acs.merge(cmp, how='left', on=['geo_id']); print(f'ACS merged with complaints is {len(final)}')
final = final.merge(opendata, how='left', on=['geo_id']); print(f'Opendata merged with complaints, ACS had {len(final)} rows')
final = final.merge(stops, how='left', on=['geo_id']); print(f'Police stops merged with Opendata, complaints, ACS had {len(final)} rows')
final = final.merge(geo, how='left', on=['geo_id']); print(f'Geographic data merged with Opendata, complaints, ACS, stops had {len(final)} rows')

# Exporting the merged file
final.head()

final.to_csv(here('./data/Combined.csv'), index=False)
